requirements.txt
# --- LangChain Core & Extensions ---
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-huggingface==0.3.0
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8

# --- Vector Store (Local Qdrant) ---
qdrant-client==1.14.3

# --- Embeddings & LLMs ---
transformers==4.52.4
sentence-transformers==4.1.0
torch==2.7.1

# --- Reranking (Cross-Encoder) ---
scikit-learn==1.7.0
numpy==2.3.1

# --- PDF Parsing ---
pdfminer.six==20250506
PyMuPDF==1.26.1

# --- API ---
fastapi==0.115.13
uvicorn==0.34.3

# --- Config ---
pyyaml==6.0.2
pydantic-settings==2.10.0

# --- Evaluation ---
sacrebleu==2.5.1
rouge-score==0.1.2
bert-score==0.3.13
pandas==2.3.0
requests==2.32.4

# --- Utilities ---
tqdm==4.67.1